third_party/blink/web_tests/fast/css/hover-after-drag-expected.txt

Test that mouse state is correctly reset after a drag. To run this test, use a mouse to drag the link "Drag me" around. It should change from blue to red while being dragged, and red to green once the drag ends by releasing the mouse button.

Drag me
PASS successfullyParsed is true

TEST COMPLETE

PASS window.getComputedStyle(document.getElementById("drag")).color is "rgb(0, 0, 255)"
PASS window.getComputedStyle(document.getElementById("drag")).color is "rgb(255, 0, 0)"
PASS window.getComputedStyle(document.getElementById("drag")).color is "rgb(0, 128, 0)"
